Rename Class_Schedule.py module to scheduling.py to better match pep guidelines
